requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
sqlalchemy==2.0.10
python-dotenv==1.0.0 
psycopg2==2.9.6
email-validator==2.0.0
pydantic[email]==1.10.2
pyjwt==2.6.0
pytest==7.3.1
pytest-mock==3.10.0
python-dotnev==1.0.0
pytest-dotenv==0.5.2
